Completed Checkmate case, debugged issues in 'rules' and 'main'
Description

Issue with rook not being registered when computing it's movement after using new convention to signal moved rook for castling detection. Ran into infinite recursion issue when fixing issue of not accounting for when king is moving to a space where it's in check, fixed by creating a small change to ignore king movements for checkmate after checking that the king cannot move. Solved issue of checkmate being toggled when a piece is being moved; as when this happens, the piece is technically not on the board, so within the countinuous while loop in main, it detects that the king is in checkmate while a piece is being moved. Added another toggable variable to prevent checkmate detection when a piece is being moved