Optimized all files
Determined that because the pieces imaes use an alpha channel, I require the method convert_alpha(). Created a dictionary constant containing the paths of all the images for the pieces. Changed while loops to for loops for more pythonic approach. Overall, managed to eliminate 240 lines of code while maintaining key aspects and improving time complexity