Finished en passant
Made three new methods, one checking explicity for a pawn moving two spaces in the first turn, returning true if so. The second is to determine the coordinates of the pieces that can perform en passant. If these coordinates are the starting coordinates in the move function, then the LH or RH enpassant is computed into the total legal moves set. lastly is a function responsible for removing the captured piece from en passant, as it the pawn will not move to the captured pawns space, meaning it must be manually moved through a new function