Completed dialouge window
In the main function, created two dictionary constants, one for the numerical represenation of each piece and the other for the letters representing the board (ie a, b, c, etc). These are used for displaying how the pieces move and the game state. Began automatic text resize but have to wait for access to secondary monitor to test results. Now beginning tooltips, the component of the game before beginning computer generated moves.