Completed part of random move generation, changed addText
Got the random move generation to work for simply moving the pieces. But there isn't any checks for check, checkmate and special moves including en passant and castling do not work either. Also made the addText method recursive so that when it needs to expand the list further, it will only eliminate the last entry, then append an empty string. Then it will recursively call itself, adding the text to the index of the empty string.