Fixed castling and check bug
There was a bug where if the king is in check, you could still move the king into a position where the king is in check. Fixed this by changing the enemy moves list to tempBoard, so if the king is in check in tempBoard, then the move is not possible. Also the program allowed for castling even while in check, so I added if the old position of the king is within the enemy moves list, it cannot castle. Also added if the new position when the king castles is in this list, then it too cannot castle.