Fixed checkmate issue
Major issues for computing checkmate during computer mode. Because the board flips only in two player mode, must compute the indecies of pawn movements differently than opposed to computer mode, where the board stays still. Furthermore, have to change opposite variable based on the colour of the computer to prevent the indecies from going out of range.Also fixed issue where if the king cannot move and is in check, then it is checkmate. Before it just went to an infinite while loop trying to figure out a legal move, even though there isn't any. Just added a return statement if the above condition is true.