Fixed checkmate and minor en passant issue
Checkmate was not being checked properly for certain instances as it didn't account to remove the piece from the board when checking the possible moves during check. In doing so, there were duplicate pieces when checking for checkmate, which under specific circumstances didn't determine checkmate as required. Changed this by moving the main component of computerMove to a seperate function. Then I reused this function for checkmate, and actually created a new board with the correct piece moved to that position. This way, it checks for all possible moves with the correct piece in the place, ensuring validility. En passant wasn't working for when player was black. Turns out just needed to seperate the cases into three states: 2 player mode, when its the computers turn or when its the players turn.