Added priority capture
Added an evaluation for each piece type. The computer will now capture according to priority of the evaluation, meaning it will capture higher evaluated pieces first (ie queen, rook). It will also capture a piece whenever it can. If it cannot capture a piece, it will move randomly. Encountered an issue where if you click a piece that is moving many times and it gets captured, it may duplicate. I believe it might be the dragging animation that is causing this error.