Fixed rapid click issue
It was because I was detecting the piece that was under the mouse before the computer move. In doing so, the old coordinates were already compute so if you clicked very fast, then the just captured piece would reappear. Also fixed an issue that pawns could not trigger check detection. Just changed the opposite parameter to 1 for only computer.