Found and fixed checkmate error
During checkmate check, I was manipulating the board incorrectly. When the piece being inspected captures an enemy piece, it didn't correctly revert back to the original state of the board, instead making the captured piece dissapear. Because this is a very speicifc scenario, I didn't manage to detect it untill yesterday night. Also found out that if you toggled 'Raised Exceptions' in the 'Breakpoint' menu when debugging, the exceptions from a try and catch statement will be shown and stop the program.