Fixed minor issues
In checkmate, the temporary board used to determine checkmate was copying the wrong board state. It was copying the board state from the previous check and not from the original board, meaning that there were duplicate pieces within the baord. Just changed the board that it was copying inside the for loop. Also had an issue where if promoting a piece lead to checkmate, the dialouge window didn't update. Just drew the dialouge window everytime that promotion happened. I'm a bit stuck on how to return the new coordinates of the compared moves within minimax.