Added move returns in minimax
minimax function can now return moves based on it
's predictions. Increased the value of each piece by 10 times to be greater than the positioning of the pieces. Next steps are to include beta pruning, a more complex piece square table and maybe teach it openings once everything is done on time