Digital Signature
Added modifying the properties of the executable file using a version file template found from windows. Using Microsoft's SignTool, added a digital signature to the file. Currently trying to fix an issue where the wrong value for the variable 'opposite' is being returned to the pawnMove function, causing the function to compute the pawn movements from the wrong perspective, causing an index error. So far only happens when the search tree reaches a depth of 3 or greater, so will either lower the depth and see what happens or use try catch statements to debug the issue. Also had to create a new branch and deleted the old default branch because I accidentally put the password to my digital certificate in Package.py....