Package optimization
Made a new function that reads and rewrites files automatically. Changed the main component of the code into the method 'main', using __name__ == '__main__'. Also included a prompt that will ask the user for the certificate password, that is protected from the getpass method. Also included a prompt to ask the user wether or not to copy Chess.exe to desktop