Fixed horizontal line wrapping
My first version of horizontal line wrapping wasn't exactly dynamic so it didn't work as desired. This version will do so for all screen sizes. Also added in the to check wether or not the king is in check directly after promotion, something I forgot to do before. Game is pretty much done, as any other work will be to improve upon the AI.